# apps/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Afframe Public API",
    "version": "1.0.0",
    "description": "Public API for the Afframe accounting platform. Authenticate with an API key as a bearer token in the form `affk_live_…`. Sandbox keys (`affk_test_…`) are not issued yet. Errors are returned in a Plaid-shape envelope; rate limits are surfaced via IETF `RateLimit-*` headers.\n\nService status: [status.afframe.com](https://status.afframe.com) (programmatic at `GET /v1/status`).",
    "contact": {
      "name": "Afframe support",
      "email": "support@afframe.com",
      "url": "mailto:support@afframe.com"
    },
    "license": {
      "name": "Proprietary — see Terms of Service",
      "url": "https://afframe.com/terms"
    }
  },
  "servers": [
    {
      "url": "https://api.afframe.com",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "API key in the form `affk_live_…`. Send as `Authorization: Bearer <key>`. Sandbox keys (`affk_test_…`) are not issued yet — see the sandbox roadmap in the docs."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "API key missing, malformed, revoked, or pointing at a different environment than the host.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "error_type": "UNAUTHORIZED",
                "message": "Missing or invalid API key.",
                "requestId": "req_1f5a8c6e91d240bdbe18d4e07a3f9c14"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "API key authenticated but lacks the scope required for this operation, or the target resource belongs to a different tenant.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "error": {
                "code": "forbidden",
                "error_type": "FORBIDDEN",
                "message": "The API key is not allowed to perform this action.",
                "requestId": "req_1f5a8c6e91d240bdbe18d4e07a3f9c14"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource referenced by the request URL does not exist (or does not exist within the authenticated tenant — Afframe never leaks cross-tenant existence).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "error": {
                "code": "not_found",
                "error_type": "NOT_FOUND",
                "message": "Resource not found.",
                "requestId": "req_1f5a8c6e91d240bdbe18d4e07a3f9c14"
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency conflict, optimistic-concurrency mismatch, or business-rule conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "error": {
                "code": "conflict",
                "error_type": "CONFLICT",
                "message": "Resource is in a state that conflicts with the request.",
                "requestId": "req_1f5a8c6e91d240bdbe18d4e07a3f9c14"
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request body parsed but failed schema validation. The `details` array carries one entry per offending field.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "error": {
                "code": "validation_error",
                "error_type": "VALIDATION",
                "message": "Request body failed validation.",
                "requestId": "req_1f5a8c6e91d240bdbe18d4e07a3f9c14",
                "details": [
                  {
                    "path": "$.legalName",
                    "code": "required",
                    "message": "legalName is required."
                  }
                ]
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Caller exceeded the API-key rate limit. The response carries IETF `RateLimit-*` headers indicating the reset window.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "error": {
                "code": "rate_limited",
                "error_type": "RATE_LIMITED",
                "message": "Too many requests. See the RateLimit-* headers for the reset window.",
                "requestId": "req_1f5a8c6e91d240bdbe18d4e07a3f9c14"
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable error code. SDKs map this to typed error classes; do not switch on `message`. New codes may be added over time (MINOR change) — treat unknown codes by their `error_type` family.",
                "enum": [
                  "bad_request",
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "conflict",
                  "idempotency_conflict",
                  "stale_resource",
                  "feature_not_enabled",
                  "payload_too_large",
                  "validation_error",
                  "rate_limited",
                  "internal_error"
                ]
              },
              "error_type": {
                "type": "string",
                "description": "Plaid-shape family (INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, VALIDATION, RATE_LIMITED, INTERNAL, …)."
              },
              "message": {
                "type": "string",
                "description": "Developer-facing message. Safe to log."
              },
              "display_message": {
                "type": "string",
                "description": "Optional end-user-facing message. Present when the cause is safe to surface in a UI; absent for 5xx / unknown errors."
              },
              "documentation_url": {
                "type": "string",
                "format": "uri",
                "description": "Optional deep-link into a hosted error registry. Reserved for future use; the api does not emit this field today."
              },
              "requestId": {
                "type": "string",
                "description": "Echoes the `X-Request-Id` header for support tickets."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "code",
                    "message"
                  ]
                },
                "description": "Field-level error breakdown. Reserved for `422 validation_error`; not emitted today."
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Plaid-shape error envelope. Every non-2xx response from `/v1/*` is wrapped in this object."
      },
...
      "CreateFeedbackRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "bug",
              "request",
              "issue",
              "question"
            ],
            "description": "Feedback category. `bug` = something broken; `request` = new feature ask; `issue` = process / UX / docs problem; `question` = support question that didn't fit the FAQ.",
            "example": "bug"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000,
            "description": "Free-form feedback text. 1–4000 characters. Markdown allowed but not rendered — the Linear issue treats this as plain text.",
            "example": "The /v1/organization endpoint returned a 500 when my API key was created in the last 60 seconds. Retried after a minute and it worked."
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "description": "Optional contact email. If provided, support will reply here; otherwise the response is filed without a return path. Already-authenticated keys carry an org contact so this is purely for users who want a direct reply.",
            "example": "dev@partner.example"
          },
          "context": {
            "type": "object",
            "properties": {
              "page": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "maxLength": 2048
                  },
                  "pathname": {
                    "type": "string",
                    "maxLength": 512
                  },
                  "title": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 500
                  },
                  "locale": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 16
                  },
                  "theme": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "enum": [
                      "light",
                      "dark",
                      "system",
                      null
                    ]
                  },
                  "referrer": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 2048
                  }
                },
                "required": [
                  "url",
                  "pathname"
                ],
                "description": "Page the report was filed from."
              },
              "scope": {
                "type": "object",
                "properties": {
                  "org_slug": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "reporter_id": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "reporter_email": {
                    "type": "string",
                    "maxLength": 320
                  }
                },
                "description": "App scope + server-resolved reporter identity. Advisory only — never used for authorization."
              },
              "element": {
                "type": "object",
                "properties": {
                  "tag": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "data_slot": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 128
                  },
                  "role": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 64
                  },
                  "id": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 128
                  },
                  "classes": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 500
                  },
                  "text": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "dom_path": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "bounding_rect": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "properties": {
                      "top": {
                        "type": "number"
                      },
                      "left": {
                        "type": "number"
                      },
                      "width": {
                        "type": "number"
                      },
                      "height": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "top",
                      "left",
                      "width",
                      "height"
                    ]
                  }
                },
                "required": [
                  "tag"
                ],
                "description": "Right-clicked element descriptor."
              },
              "selection": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 4000
                  },
                  "html": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 4000
                  },
                  "rect": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "properties": {
                      "top": {
                        "type": "number"
                      },
                      "left": {
                        "type": "number"
                      },
                      "width": {
                        "type": "number"
                      },
                      "height": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "top",
                      "left",
                      "width",
                      "height"
                    ]
                  }
                },
                "description": "Active text selection, if any."
              },
              "surrounding": {
                "type": "object",
                "properties": {
                  "nearest_heading": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 300
                  },
                  "inferred_block": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 128
                  },
                  "nearby_text": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "description": "Nearby DOM text for locating the report."
              },
              "viewport": {
                "type": "object",
                "properties": {
                  "width": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 20000
                  },
                  "height": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 20000
                  },
                  "scroll_y": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000
                  },
                  "device_pixel_ratio": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 8
                  }
                },
                "required": [
                  "width",
                  "height",
                  "scroll_y",
                  "device_pixel_ratio"
                ],
                "description": "Viewport dimensions at capture time."
              },
              "client": {
                "type": "object",
                "properties": {
                  "user_agent": {
                    "type": "string",
                    "maxLength": 800
                  },
                  "platform": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 128
                  },
                  "language": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 32
                  },
                  "timezone": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 64
                  },
                  "online": {
                    "type": "boolean"
                  },
                  "prefers_dark": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "user_agent"
                ],
                "description": "Reporter's browser/runtime environment."
              }
            },
            "description": "Optional in-app capture context (page, element, viewport, client). Public callers omit it; the in-app reporter attaches it. Folded into the Linear issue when present."
          }
        },
        "required": [
          "type",
          "message"
        ],
        "description": "Partner feedback submission. Idempotency is not enforced — duplicate submissions create duplicate Linear issues. Clients should rate-limit themselves to one submission per minute."
      },
      "CreateFeedbackResponse": {
        "type": "object",
        "properties": {
          "received": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Always `true` on a 2xx response. Confirms the api accepted the submission for downstream dispatch (email + Linear issue)."
          },
          "referenceId": {
            "type": "string",
            "description": "Opaque submission reference. Quote this in any follow-up email — it links the support reply back to the Linear issue.",
            "example": "fb_2ZdAk5x"
          }
        },
        "required": [
          "received",
          "referenceId"
        ],
        "description": "Confirmation that feedback was accepted. Acceptance does NOT guarantee a reply — `question` submissions get one, the others are reviewed without acknowledgement unless an email is provided."
      },
...
    },
...
  },
  "paths": {
...
    "/v1/feedback": {
      "post": {
        "operationId": "createFeedback",
        "summary": "Send feedback",
        "description": "Submit a bug report, feature request, process issue, or question. The api forwards every submission to `support+feedback@afframe.com` and creates a Linear issue tagged with the feedback type. Public — no API key required.",
        "tags": [
          "Feedback"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback accepted for downstream dispatch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFeedbackResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    }
  },
...
}